Use Seaborn (https://seaborn.pydata.org) AMP abstract plots.
See plot_amp.py --help or comments at the top of plot_amp.py for usage.

This should give the same trend lines (track the median, error bars for the
middle 50% interval) as before, but with more options for tweaking and
an option to overlay trend lines.

Drawing median values directly on the plots should still be doable, but I
haven't undertaken to do that yet.

This also now loads data into a Pandas dataframe, which may be useful to pull
out for other analysis of model performance.